+++ src/Resources/Wacc/index.ts
@@ -4,6 +4,8 @@ import WACCReturnOrders from './ReturnOrders/ReturnOrders';
 import WACCSearch from './Search/Search';
 import WACCUsers from './Users/Users';
 import WACCGifts from './Gifts/Gifts';
+import WACCProducts from './Products/Products';
+import WACCProductAssociations from './ProductAssociations/ProductAssociations';
 
 export interface WaccResources {
     orders: WACCOrders;
@@ -11,6 +13,8 @@ export interface WaccResources {
     search: WACCSearch;
     users: WACCUsers;
     gifts: WACCGifts;
+    products: WACCProducts;
+    productAssociations: WACCProductAssociations;
 }
 
 export function loadWaccResources(client: Client): WaccResources {
@@ -20,5 +24,7 @@ export function loadWaccResources(client: Client): WaccResources {
         search: new WACCSearch(client),
         users: new WACCUsers(client),
         gifts: new WACCGifts(client),
+        products: new WACCProducts(client),
+        productAssociations: new WACCProductAssociations(client),
     };
 }